refactor(capabilities): align the small structure deviations in rpc, render, lifecycle, ui, gemini and shared
Six mechanical, behaviour-preserving structure alignments in
crates/aether-capabilities/src, per the ADR-0122/0123 cap anatomy:

- rpc: gate rpc/server's connection module on feature = "runtime"
  instead of not(wasm), matching its sole consumer (runtime.rs).
- render: move the mid-file test module to end-of-file under
  cfg(all(test, feature = "render-runtime")) and drop the file-root
  use runtime::* glob (with its wildcard_imports/unused_imports allows).
- lifecycle: relocate the runtime-gated inspect impl and the test_cap /
  tick_start_graph_cap fixtures from mod.rs into runtime/mod.rs and
  drop the file-root use runtime::* glob.
- ui: fold the UiCapability identity ZST into ui/mod.rs, delete
  widgets.rs, add the missing module header on kinds.rs, and replace the
  pub use core::iter::once / core::mem::swap re-export shim in
  runtime.rs with plain private imports.
- gemini: fold the GeminiCapability identity ZST from capability.rs
  into gemini/mod.rs and delete the file, matching the sibling caps.
- shared: rename shared/contentgen/shared.rs to transport.rs (retiring
  the crate::shared::contentgen::shared path stutter) and update its
  four consumers.

Closes #2478
